Handle nullable lists
... and a few other potential NPEs (with some drive-by clean-up, which
I can split out if necessary).

+++ interfacemanager/interfacemanager-impl/src/main/java/org/opendaylight/genius/interfacemanager/commons/InterfaceManagerCommonUtils.java
@@ -756,7 +756,7 @@ public final class InterfaceManagerCommonUtils {
                 .build();
         tx.delete(intfid);
 
-        if (interfaceNameEntries.size() <= 1) {
+        if (interfaceNameEntries != null && interfaceNameEntries.size() <= 1) {
             tx.delete(dpnToInterfaceId);
         }
     }
